refactor(diagnose): extract InvestigateLandingView from DiagnoseView
The Investigate landing surface and the focused-endpoint detail surface
are conceptually distinct UX panels that previously shared a single
1800-line component. Splits the landing into its own
\`InvestigateLandingView.svelte\` so both surfaces stay independently
auditable; DiagnoseView keeps ownership of the focused-endpoint detail.

The extracted component pulls its inputs directly from the project
stores rather than receiving props, matching the pattern used by the
other view-level components. ~170 lines of markup + derived state and
~250 lines of landing-only CSS leave DiagnoseView. No user-visible
behaviour change; the existing empty-state copy moves with the markup
and gains an inline "Add endpoint" CTA so the landing's empty card is
both informative and actionable.

Closes Arc C cohesiveness gap B3. The larger EndpointDetailView
extraction is intentionally deferred — the detail surface owns a much
larger reactive-state surface (correlation matrix, network context,
bufferbloat, distribution interpretation) and deserves its own scoped
refactor PR.
